Extract `Interns` to a new module.
